requirements.txt
setuptools~=63.4.1
sqlmodel~=0.0.6
pydantic~=1.9.2
SQLAlchemy~=1.4.40
croniter~=1.3.5
python-snap7~=1.2
pycomm3~=1.2.10
paho-mqtt~=1.6.1